+++ CHANGELOG.md
@@ -76,6 +76,8 @@
 8. 增加了对于QA_Performance插件中先卖后买算差价的支持
 9. QAMarket=QAShipaneBroker 支持通达信模拟客户端
 10. QAIndicator 增加Talib的形态识别指标
+11. 增加了qarun.exe 直接运行策略,实时输出
+12. QAWEB 增加了command/run
 
 不兼容修改: